+++ src/view/empty-list-view.js
