+++ src/view/failed-load-view.js
@@ -1,10 +1,10 @@
 import { createElement } from '../render.js';
 
-const createLoadingTemplate = () => '<p class="trip-events__msg">Failed to load latest route information</p>';
+const createFailedLoadTemplate = () => '<p class="trip-events__msg">Failed to load latest route information</p>';
 
-export default class LoadingView {
+export default class FailedLoadView {
   getTemplate() {
-    return createLoadingTemplate();
+    return createFailedLoadTemplate();
   }
 
   getElement() {
